Add new column for tracking management_enabled
Added a new column to ComplianceStatus that tracks whether
the consumer has an entitlement that sets management_enabled.

This value is populated when a consumer event is processed.

This approach was a lot easier than having the query determine
it and avoids another subquery. The data does not change once
received and stored, so this is Ok.