1199443: Fix entitlement validity dates for temp entitlements.
This patch corrects the isValidOnDate() method to account for the new
unmapped guest entitlement functionality.  The isValidOnDate() method
now examines the pool attributes to determine if an entitlement is for
unmapped guests and if so calculates the end date based on the consumer
creation date rather than using the pool end date.