Fix issues with product/content manifest import.
Still work to be done here, product and content should be imported in the same
place. We also need to clarify when things get saved, we can't do it during
import because we need to look for changes vs the old version during refresh,
so content storage should probably move to CandlepinPoolManager as well. Sadly
this isn't consistent with import of other non-subscription objects.