Add some curator.refresh() calls after creating objects.
*Importer.store() methods get a curator.refresh(new_object)
added after the create.

This seems to speed content/product import up, from
about 7 secs for content, and about 2 mins for products, to
about 3 secs for content, and about 1 sec (total) for products.

Refresh is still pretty slow, but on a big import, that shaves
off about 2mins from the total.

Why?  I'm not entirely sure. Theory was that hibernate was
calculating db transaction 'dirtinesss' on every insert or
select, so added refresh to try to get the transaction
caught up.