Consolidate request and response logging into one log statement.
Previously the components of a request and/or response were spread
across multiple log statements.  One at INFO level would log the verb
and requested URL (or the response code) and one at DEBUG level would
log the headers and body.  Splitting what should logically be a single
unit across multiple statements made reading the logs harder than
necessary.

This patch consolidates an entire request and/or response into a single
statement.  If the logger is set to INFO, it writes a short statement
with the verb and URL combination (or response code).  If the logger is
set to DEBUG, it skips the aforementioned INFO statement and writes a
longer statement with the headers and body.