Do not process messages already handled by gutterball
If qpidd is restarted without gutterball, messages will
be resent. Gutterball will now store a message ID on each
event and will make sure that an Event with the incoming
message ID does not exist, before processing.

There is no way to determine the originating message ID
of already persisted Events, so these will be set to
"UNKNOWN" on DB upgrade.