Retain backward compatability with clients listing pools.
New model changes altered the pool json to the point where clients in the wild
would break when trying to parse provided product data. The discrepancy between
the meaning of id vs productId was too much to try to hack around. Instead
we'll hide the much more massive serialization of the actual product objects
(and all their content), and instead serialize a DTO that looks very much like
the old response data as "providedProducts" and "derivedProvidedProducts".

This has the added benefits of helping us parse legacy manifests (i.e. the
transient fields added to pool for this incoming data), and returning the rules
closer to the state they were in before, potentially helping us *not* require a
major rules version update. (blocking their use on old servers)