Use FlushMode=COMMIT for manifest imports
Add a setFlushMode() to AbstractHibernateCurator.

Use above to set the FlushModeType for the Importer.importObjects
portion of a manifest import to "COMMIT", instead of whatever
the default chooses (AUTO). With a large transaction and
flushMode=AUTO, any query will end up needing to check any
object in the transaction for updates, and spend a lot of time
doing that. FlushMode.COMMIT flushes the transaction on
commit, but mainly it turns off AUTO mode and prevents the large
number of dirty checks on objects.

For something like the manifest referenced in
https://bugzilla.redhat.com/show_bug.cgi?id=1193803, the manifest
import was previously taking ~4-5 mins, now takes about ~8 secs.