Add the Maven assembly plugin to generated POM files.
The Maven assembly plugin creates a tarball of the entire project which
we need when building the RPM since the Maven generated artifacts don't
include the license, conf files, etc.

The buildr pom task now adds inserts the assembly plugin and adds the
ability to insert other plugins with Procs defined in the project
definition.