setup-dev for rhel7
A few changes here:

1) supervisor is only in epel, so setup a epel repo.

2) There are a few packages that need update since the rhel7 image,
one of which is subscription-manager's container support, so add
a yum update -y to get it.

3). More locale weirdness, likely better fix for this.

4). Install 'yum-utils' so we can get 'yum-config-manager', since
'subscription-manager repos' doesn't work in container mode. We need
to enable the 'rhel-7-server-optional-rpms' repo to get 'ruby-devel'
so we can build the bundler deps we need.

5. Then install the packages.

6. Oh, this also does a 'git pull' after setting up the remotes,
so all the pr refs exist in the image.