Fix cross-org import bug.
If two orgs have a manifest with a the same product that has changed at all
since the manifests were generated, import in the second org will cause the
first org to wipe out all but one provided product on the first org's
subscription. If the first org then re-imports their manifest to correct the
problem, it will clobber the second org's subscription provided products.

Situation only resolves if both orgs use an updated manifest with the same
version of product.

Caused by an issue in hibernate where a criteria query on children of a class
will only populate the child that matched the query on the object once loaded
in memory. Switching to HQL corrects the issue and loads the full collection.