Restore serialization of duplicated pool.product data.
Because we use the same serialization for both the API and rules, problems were
being encountered where when serializing the response from compliance JS,
entitlement.pool.product* was null/empty. As a result anything that relied on
the response could not properly process pool attributes leading to problems
with autobind attempting to heal with existing entitlements.

Will continue serializing duplicated data for time being.