Include derived modifier content in distributor ent cert
In the case of a distributor, derived products were not being
considered as entitled products. This would cause any content
specifying modified products to not be included in the
distributor's entitlement certificate, preventing content
sync downstream.

Candlepin now considers all derived products as entitled products
for a distributor. Standard rules governing inclusion of modifier
content still apply, but in the case of a distributor, modified
products specified on content can now be a derived/derived provided
product.

This means that in order for modifier content to be included in
the entitlement cert, the consumer must be entitled to the
modified product in one of the following ways:
- consumer currently has an entitlement providing the modified product
- the pool from which the entitlement cert is being generated from
  provides the modified product.
- the pool from which the entitlement cert is being generated from
  provides the modified product via derived products.