Fix cross-compile issue with ConcurrentHashMap.
The Java 1.7 ConcurrentHashMap#keySet() returns a Set<K> while the Java
1.8 version returns ConcurrentHashMap.KeySetView<K>.  Since Java allows an
overridden method to return a sub-type of the parent method's return
type, no compilation are raised.  When we run the code on Java 1.7,
however, it immediately fails since the ConcurrentHashMap.KeySetView
class is not available.

One solution is to compile against a 1.7 bootstrap classpath, but since
1.7 is not packaged for recent versions of Fedora, we will work around
the issue by upcasting ConcurrentHashMaps to just Maps.

See https://gist.github.com/AlainODea/1375759b8720a3f9f094 for a
detailed explanation.