Fix EL7 docker rpm container test.
Addition of set -e introduced problems in the startup.sh script, as the cpsetup
command does error out on EL7 because the service it not yet running, thus we
can't check status and initialize. Previously this would fail and continue in
the script, starting supervisor.

Instead we'll keep the set -e, but add a new cpsetup option that skips service
stop/restart/init, and eliminates the failed command during startup.

Also added wget calls to fetch the latest cpsetup and cpdb for this base
container used for rpm installation tests. This isn't used by the more official
tagged containers with a build in them, only for the base image that doesn't
yet have cp installed.