Reduce database chatter in GET /status.
This looks like it might be the largest chunk of processing our app does in
prod. Client makes ample use of this, but IT also uses it as a health check.

First step drops querying of rules version and source, instead we set these on
the provider singleton whenever we recompile rules. (i.e. they changed) The two
places they can change (rules upload and import) are updated to immediately
recompile rather than wait for someone to attempt to use them.

Ideally we'd have dropped all db connectivity from the class, but unfortunately
we need to know if the db loses it's connection, so IT can drop the node from
rotation automatically. Instead, we'll reduce the query dramatically and just
query the updated timestamp.

A nice future enhancement would be to catch db errors (all of them somewhere)
and set some kind of flag once we see any.

Drop custom creation of a principal for compiling rules.

This is a very old bit of code, can't find any reason why we need a principal
to lookup rules, and candlepin seems fine without it now.

Make compileRules public for places we know we want to.