Storage band implementation
Implemented storage band counting. Each entitlement
will equate to 1TB of coverage and marketing products
will use the multiplier such that sub quantity purchased
* multiplier will yeild pool quantity (total capacity in TB).

For example:
pool_quanity = sub_qty_bought * (storage_band_attr * multiplier)
             = 1 * (1 * 256)
             = 256

The resulting pool will have a quantity of 256, which can
cover 256TB of storage.

The rules will calculate a system's coverage based on a
band.storage.usage fact.

Consider a system is using 128TB:
- a quantity of 128 entitlements will cover the system (green).
- a quantity < 128 entitlements will partially cover the
  system (yellow)
- no entitlements (red)

Marketing Product (requirements)
---------------------------------
The marketing product should define a multiplier that is equal
to the total capacity in TBs. It must be stackable and must
define the storage_band attribute (should be set to 1).

- stacking_id: "product_stack_id"
- mutli-entitlement: "yes"
- storage_band: 1

Consumer Requirements
-------------------------
The consumer must have the band.storage.usage fact set to
the total TB that it is using.

NOTE: If this fact is not set, the rules will assume a value
      of 1TB.

- band.storage.usage: 128