Revert consumer/hypervisor last check-in tables.
This was intended to help reduce lock wait timeouts by removing contention on
cp_consumer.lastcheckin. However we later discovered the performance impact was
just too great as the size of the tables grew and unfortunately the work has to
be reverted.

This however was only to fix a symptom, long running transactions were the
underlying cause, and we'll have to tackle this once that can be solved. Recent
changes (virt-who) may help with this.

Upgrade script is included to revert the values back to the max in the db at the time.

Quartz will also have to be updated to remove refs to the jobs that are being
removed. In IT's case this will need to be done manually as their quartz tables
are in a separate database.