1256926: Fixed Pool match filter query
When using match parameters while listing available pools
hibernate would return a partially initialized list of
pool objects when matching on a collection of sub objects.

For example, if the match filter matched on 1 of 4 provided
products, only the matched provided product would be loaded
in the Pool's collection. Only a hibernate refresh would
bring the remainder in after the initial lookup.

This patch changes the query to use correlated subqueries
instead of left outer joins, which works around the partial
collection issue.

I tried implementing a unit test for the curator but
wasn't able to reproduce. I have a hunch that HSQLDB
may be dealing with joins a little differently than
postgres. A spec test easily reproduced this issue.