Small fix to performance script
Added owner_key parameter when creating products.

******
 NOTE
******

Since moving to per org products this script is
now out of date. Since it relies on existing
products from our test data, it will only run
successfully against existing orgs since the
installed products we are using will not exist
when running against a new org.

In the future, we should revamp this script to
create the set of products that we will be
testing with.

******