Limit the number of binds per consumer that can run concurrently.
Binds can take a long time due to the calculation of compliance rules.
If we don't throttle the number of concurrent bind jobs, it is likely
that some of the jobs will time out waiting to acquire a write-lock in
the database.  This commit changes Pinsetter to only run a configurable
number of bind jobs concurrently.  Any jobs above the threshold are
temporarily paused.