Fixed Optimistic Lock issue with ConsumerComplianceJob
If a consumer was updated or deleted before the job was run
an OptimisticLockException was thrown and the job would fail.
Losing a compliance can be problematic as the consumer
status would be out of date.

The OLE would occur if the consumer was updated/deleted after
it was looked up and before it was locked.

This PR locks the consumer isntance on lookup, eliminating the
gap between fetch and lock.

If the Consumer does not exist, the job will now finish
successfully with an appropriate result message stating that
the targeted consumer could no longer be found.