Swagger Converters from 1.5.7
This commit doesn't include any behavior changes to Swagger. It is adding
Swagger converters but with code that is copied out from Swagger 1.5.7
ModelResolver class.  There are some small changes we need to make in subsequent
commits to the default converters so that Swagger generates models consistent
with our serialization style. Because the changes couldn't be done
incrementally, I decided to take this copy/modify approach.