Add missing @Consumes annotations.
Also add warning when a @Consumes is set to the */* media type but
actually reads the request body.  Currently, Candlepin only consumes
JSON so methods should be specific about that instead of using the
wildcard type as a shortcut.

It is appropriate, however, to specify the wildcard type in @Consumes
when the method does not read from the request body at all.