1337906: Stack derived pool cleanup
In some code paths, it may be possible that deletion of entitlements
might not succeed (DELETE statement is not issued to the database).
The underlying issue here is with collection Pool.entitlements. Because
this collection has cascade set to PERSIST, one must be very careful
when deleting entitlement to also remove it from the collection.
Otherwise, Hibernate will un-schedule the deletion.