1345765 - Concurrent content demotions cause 400
Under concurrent demotion requests, we were getting StaleStateException
from hibernate indicating that a request is trying to delete already
deleted instance. Also, the demotion of content was using one
transaction per content demotion, so request could demote some content,
then throw exception and never get into the part of the code that is
responsible for regeneration of environment certs.

In this commit I introduce RdbmsExceptionTraslator that can help to
determine a reason for a Hibernate Exception. Also I use batch
operation for database operation (demotion of content) to mitigate a
possibility of partial demotion.