1345766: Concurrent content promotion cause 400
Similar to 1345765. This method was designed in the past to throw 409 in
case when a content has already been promoted to an environment. This
wasn't working sometimes under concurrent access. Also the method allows
to pass in a set of content ids to promote but it is possible that
method just partially promotes a subset of those ids.

Adding transactionality and exception handling.