1350771: ExclusiveLock held by Candlepin
Removing foreign key constraints on checkin tables to mitigate a
posisibility of lock waits during GET requests.

When a long running request acquires X lock on a consumer, then any
other requests by the consumer (even GET) will get blocked. This is
because we update consumer checkin tables that have foreign key to
consumer table. This behavior is specific to Postgres 9.2, newer
Postgres has a new feature that is less strict to updates to foreign
tables.