1354633: Cleanup stray ueber objects
An old bug that caused the ueber cert pools
to be deleted was fixed but did not correct
the state of any owners that were affected.
All required ueber objects (sub, product,
content, consumer) were not cleaned up.

Subsequent attempts to get or create a new
debug cert would fail because the data was
in a new state.

This patch adds a new liquibase changeset to
clean up the data. This patch plus the old
bug fix should make sure that this doesn't
happen again.

Another adjustment on all dbs