1350455: Improve case insensitive search
Changing strategy for case insensitive searches. In the past we
used ignoreCase() and lower() to perform case insensitive search.
That strategy was performing poorly in the absence of an database
index. Adding such index was not possible in database agnostic way
so the new strategy is to use columns suffixed '_lower'. These
columns have index created and they contain only lower case data.