Add ansible automation to optionally setup yourkit profiling
This will load all env vars beginning with "CANDLEPIN_" (case-insensitive)
and their values (less the prefix of "CANDLEPIN_") into ansible.
The idea here is that in the future if we have additional optional
configuration we'd like to add to the ansible playbooks we can simply
start using the new variable in the playbooks (checking to see if the var
is defined) and add it as an environment variable to actually use it.