Ensure debug symbols make it into compiled code.
It looks like the class files that our in our official builds do not
have debug symbols in them.  No debug symbols saves some space but also
means that Support can't install a Satellite instance and then attach to
a running Candlepin with jdwp.

I have a suspicion that MEAD sets debug=false in an environment
configuration somewhere because the debug symbols are added when I
compile locally.  Setting the values explicitly should ensure the
symbols are left in.