Reference caching of Product
Reason for these changes is to improve performance of code paths that use
Pool.providedProducts and Pool.derivedProvidedProducts. The basic idea is that
Product entities are cached in key-value cache by reference. Thus, all the
thread of Candlepin share the same instances of Product entities. Furthermore,
upon first load, the Product instances that are about to be placed in the cached
are fully hydrated (all lazy loaded collections, such as Content are loaded).
Underlying assumption, that allows us to do this change, is that Product
entities are fully immutable - so there is no chance that cached Product might
be come stale.

Most of the references to getProvidedProducts and getDerivedProvidedProducts
has been removed from the code. The lookups to provided products should be
instead made through ProductCurator. The methods on ProductCurator ensure
caching of fully hydrated Product objects.

The getters still remain on the Pool. This is necessary for some situations, for
example when a Pool is 'fabricated' using
CandlepinPoolManager.convertToMasterPool the getters are the only reasonable way
to retrieve fabricated provided products.

Attributes Pool.providedProducts and Pool.derivedProvidedProducts are still
mapped as Hibernate collection. Reason for that is, that we need those
collection when creating the association between Pool and provided products. One
can argue this is more of a technical problem. In ideal situation, in the
future, we might want to remove these Hibernate colletion mappings and we might
want to create the links between Pool and provided products using an
intermediate entity between Pool or Product. Or some other mechanism that will
allow us to remove the mapped collections.

New serialization interceptor PoolSerializer is now used for all serializations
that require Pool to have provided products (REST calls, JS rules, HornetQ,
AMQP, etc.) The strategy to serialize provided products has two steps. First,
before a Pool is serialized the serializer will fill in transient properties
that contain provided products. Then, in second step, the serialization
continues as before.

Methods provides() and providesDerived() that were on the Pool are now replaced
by ProductCurator.provides and ProductCurator.providesDerived(). The new methods
on the curator are using direct database lookups, instead of traversing
collections.