Golden Ticket
Content access certificates generated for consumers when org has indicator
The entitlment and signature blocks are stored by org and environment
The certificates contain the path for org/environment in the encoded cert but
  all of the possible content in the entitlment block
Access mode assigned directly to manifest consumer.
Must exist in the mode list on owner.
When imported, that mode is both the mode and the mode list on the downstream owner
Mode on downstream owner cannot be set on downstream owner via API.
OID for Entitlement type.
Modify content access mode in manifest consumer in script
Move 'if-modified-since' to header
Add owner key to 'O' in DN