Fail Fast
New -q switch for deploy script that will recreate queue 'allmsg' which is a
special queue for testing purposes. -a switch will also cause autoconf
to enable AMQP in candlepin config.

Docker image candlepin-base now installs qpid and starts it using supervisord.

cp-test script has new -q switch that uses guterrball configure-qpid.sh
script, to setup qpid (generate SSL keys, create event exchange) etc.

I also had to modify configure-qpid.sh to be aware of supervisorctl

You can now write spec tests that assert Qpid queue. There is a new
CandlepinQpid helper class.

When connection to Qpid goes down, we must reestablish all the JMS Objects. This
behavior is dictated by JMS specification. Also, the timeout of Qpid JMS
failover had to be lowered to be able to easily, and quickly, detect broker
connection failure.

Qpid can be in three states that Candlepin needs to distinguish. First state is
'Qpid down' which is when connection to the Qpid is not available. Second state
is 'Flow Stopped'. That situation occurs when one of the queues bound to the
event exchange is flow stopped (see Flow Control in Qpid docs). Last is normal
happy 'connected' state in which Qpid is available. These states are detected
usign QMF (Qpid Management Framework)

Candlepin can enter Suspend Mode. In this mode Candlepin will only
respond to Status Resource requests and pause all Pinsetter Jobs. It is
expected that as soon as Candlepin enters Suspend Mode, no more new
requests or jobs are started. It is still possible that some of the
ongoing requests and jobs succeed.
When Candlepin is in Suspend Mode, it returns an error that is uniform
with the rest of the Candlepin. Client than can easily use /status/mode resource
to learn more details about the reason of suspend mode.

Candlepin will quickly fail at startup if the Qpid Broker is not running and
healthy.

Candlepin will now automatically transition to/from Suspend Mode. This is
achieved by class SuspendModeTransitioner that technically uses
ScheduledExecutorService thead pool (single threaded). The transitioner
periodically runs checks for Qpid connectivity but also exposes public method
that can be used to attempt transition. The transition ends up being a no-op if
no transition is needed or ends up transitioning Candlepin from/to suspend mode.

The periocity of the transitioner is using growing delay between executions. The
idea behind it is, that after the Qpid has been down for extensive period of time,
there is a high chance it won't come up soon. A config property is in place to
set maximum bound to the delay.

There is also new config property to turn off Suspend Mode.