Suspend Mode
Candlepin can enter Suspend Mode. In this mode Candlepin will only
respond to Status Resource requests and pause all Pinsetter Jobs. It is
expected that as soon as Candlepin enters Suspend Mode, no more new
requests or jobs are started. It is still possible that some of the
ongoing requests and jobs succeed.
When Candlepin is in Suspend Mode, it should return an error that is uniform
with the rest of the Candlepin. Client than can easily use /status/mode resource
to learn more details.