Batch Revoke Performance Fix
Removed the useage of ProductCachedSerializationModule in
EventFactory while serializing pools since we use the
HATEOS JSON version of Pool when serializing for Events.

Removing this serializer produces the exact same JSON
data. No need to hydrate provided products as it is
never, and has never, been included.

On revoke entitlement, do not remove entitlements
from their pool's collection. Since the collection
is Extra lazy, hibernate will query the ents each
time an entitlement is removed from a Pool's collection.

By doing this, we are no longer maintaining runtime
consistency b/w Pools and Entitlements. The biggest
concern with doing this is that we may run into a
hibernate unscheduling problem. For example a missed
delete.

To minimize this risk, this PR removes the cascades
on the Pool.entitlements collection.