1394401: List of provided products empty on Bonus Pools
The problem is in method PoolHelper.copyProvidedProducts. The method is used to
copy the original pool to a bonus pool. The method
expects the original pool (the one being copied) to be stored in the database.
It is trying to find its provided products in the database. That fails, because
the original pool is not yet there.

Adding more defensive code into that method. First I will check whether the
pool is already persisted (Pool.id != null). When it is not persisted, I will
copy the provided products directly from the original instance.