Ueber cert cleanup round 3
Added another round of ueber cert data cleanup
that omits any of the 0954 tables. This cleanup
was added in case some of the changelog 'property'
names collided in some way and the latest queries
were not run.

History:
Property definitions in changeLogs are
not overridden when defined in multiple
changelogs.