Make initial entitlement certificate generation lazy.
This removes the Certificates from the Entitlement model and
makes the initial certificate generation lazy. This was done
because neither subscription-manager or katello are using
the certificate data from the initial bind and it will
enable us to shorten the duration of the lock that is held
for the pool record while entitlements are created.