Remove numeric limitation on engineering product IDs
Remove the limitation that IDs for engineering products may only contain
numbers. This will allow any alphanumeric ID to be used for custom
product generation. In general this will allow someone to let Candlepin
generate the IDs for custom products during creation instead of having
to generate them externally and pass them in.

The limitation on numeric IDs is being replaced with a more explicit
test of whether or not any content sets are associated with the product.