Uber certificate rewrite
A rewrite of this functionality was done to
remove the complexity from the generation
process and supporting data model used to
create the certificate.

In the new approach, uber certificates are
stored in their own table, and no longer
require 'dummy' consumers/pools/constent/product
stored in the DB in order to get a certificate
for an Owner.

We still only allow one uber cert per owner
but the cert is now directly related to the
Owner record and has no ties to any other
entity.

As part of this patch, all legacy uber cert
functionality has been removed and replaced
with the new.