Removed @Formula for  CertificateSerial.revoked
No longer use @Formula to determine if a cert
serial is revoked or not. Instead, we store the
state on the object when the referencing certificate
is revoked.

The serial's revoked state is set to true on deletion
of the associated certificate via means of a @PreRemove
callback. Using the callback to update revoked comes
with a drawback. We can not delete the record via JPQL
or the callback will not be fired. It must be deleted
via cascade, or directly via the EntityManager.

Any certificate that wishes to make use of this functionality
must extend the RevokableCertificate base class, and should
follow the deletion process above.

A liquibase changeset is also included with this patch
and when applied, will add a new revoked column to
cp_certificate_serial and will update any serial with
no association to a certificate to be revoked.